+++ tests/fail/bench-invalid-sig.rs
@@ -1,0 +1,8 @@
+// Copyright (C) 2025 Daniel Mueller <deso@posteo.net>
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+
+/// Invalid signature for `#[bench]` function.
+#[test_fork::bench]
+fn invalid_sig() {}
+
+fn main() {}