+++ tests/failures.rs
@@ -9,5 +9,11 @@ use trybuild::TestCases;
 #[test]
 fn failures() {
     let t = TestCases::new();
-    let () = t.compile_fail("tests/fail/*.rs");
+    let () = t.compile_fail("tests/fail/test-invalid-args.rs");
+    let () = t.compile_fail("tests/fail/fork-env-mut-capture.rs");
+    let () = t.compile_fail("tests/fail/fork-no-inner-test.rs");
+
+    if cfg!(all(feature = "unstable", feature = "unsound")) {
+        let () = t.compile_fail("tests/fail/bench-invalid-sig.rs");
+    }
 }