+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Introduced `#[bench] atttribute for running benchmarks in a separate
+  process
 - Introduced `#[fork]` attribute that unconditionally requires nesting
   with other `#[test]`